Add conversion of Class<?> and String to ResolvedJavaType and ResolvedJavaMethod in DynamicAccessDetectionPhase
